+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
+- Gem ActAsTaggable - Forest does not make all models taggable anymore.
 - Rails Version - Fix a recent incompatibility with Rails 5.1+. [Regression introduced in 2.4.5]
 
 ## RELEASE 2.4.7 - 2018-03-21